More refactoring to remove redundancies
Removing redundancies from DNA/Protein models
