+++ phylo/src/substitution_models/protein_models/protein_model_generics.rs
@@ -1,27 +1,10 @@
-use std::collections::HashMap;
-use std::fmt::Display;
-
 use lazy_static::lazy_static;
-use ordered_float::OrderedFloat;
 
-use crate::evolutionary_models::{EvoModelParams, EvolutionaryModel, ProteinModelType};
-use crate::likelihood::LikelihoodCostFunction;
+use crate::frequencies;
 use crate::sequences::{AMINOACIDS, GAP};
-use crate::substitution_models::{
-    FreqVector, ParsimonyModel, SubstMatrix, SubstModelInfo, SubstitutionLikelihoodCost,
-    SubstitutionModel,
-};
-use crate::{frequencies, Result, Rounding};
+use crate::substitution_models::{FreqVector, SubstMatrix};
 
-pub(crate) type ProteinSubstArray = [f64; 400];
-pub(crate) type ProteinFrequencyArray = [f64; 20];
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub enum ProteinParameter {
-    Pi,
-    Mu,
-    Lambda,
-}
+use crate::substitution_models::protein_models::{ProteinFrequencyArray, ProteinSubstArray};
 
 lazy_static! {
     pub static ref AMINOACID_INDEX: [usize; 255] = {
@@ -78,202 +61,6 @@ fn generic_protein_sets(char: u8) -> FreqVector {
         set
     } else {
         frequencies!(&[1.0 / 20.0; 20])
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct ProteinSubstParams {
-    pub(crate) model_type: ProteinModelType,
-    pub(crate) pi: FreqVector,
-}
-
-impl EvoModelParams for ProteinSubstParams {
-    type ModelType = ProteinModelType;
-    type Parameter = ProteinParameter;
-    fn new(model_type: &ProteinModelType, _: &[f64]) -> Result<Self>
-    where
-        Self: Sized,
-    {
-        Ok(Self {
-            model_type: *model_type,
-            pi: match model_type {
-                ProteinModelType::WAG => wag_pi(),
-                ProteinModelType::HIVB => hivb_pi(),
-                ProteinModelType::BLOSUM => blosum_pi(),
-                _ => unreachable!(),
-            },
-        })
-    }
-    fn parameter_definition(
-        _model_type: &ProteinModelType,
-    ) -> Vec<(&'static str, Vec<ProteinParameter>)> {
-        todo!()
-    }
-    fn get_value(&self, _param_name: &ProteinParameter) -> f64 {
-        todo!()
-    }
-    fn set_value(&mut self, _param_name: &ProteinParameter, _value: f64) {
-        todo!()
-    }
-    fn get_pi(&self) -> &FreqVector {
-        &self.pi
-    }
-    fn set_pi(&mut self, _pi: FreqVector) {
-        todo!()
-    }
-}
-
-impl Display for ProteinSubstParams {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.pi)
-    }
-}
-
-impl From<ProteinSubstParams> for Vec<f64> {
-    fn from(val: ProteinSubstParams) -> Self {
-        val.pi.as_slice().to_vec()
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ProteinSubstModel {
-    pub(crate) params: ProteinSubstParams,
-    pub(crate) q: SubstMatrix,
-}
-
-impl SubstitutionModel for ProteinSubstModel {
-    type ModelType = ProteinModelType;
-    type Params = ProteinSubstParams;
-    const N: usize = 20;
-
-    fn create(params: &ProteinSubstParams) -> ProteinSubstModel {
-        let q = match params.model_type {
-            ProteinModelType::WAG => wag_q(),
-            ProteinModelType::BLOSUM => blosum_q(),
-            ProteinModelType::HIVB => hivb_q(),
-            _ => {
-                unreachable!("Protein substitution model should have been defined by now.")
-            }
-        };
-        ProteinSubstModel {
-            q,
-            params: params.clone(),
-        }
-    }
-
-    fn new(model_type: ProteinModelType, _: &[f64]) -> Result<Self>
-    where
-        Self: Sized,
-    {
-        let params = ProteinSubstParams::new(&model_type, &[])?;
-        let mut model = ProteinSubstModel::create(&params);
-        model.normalise();
-        Ok(model)
-    }
-
-    fn index(&self) -> &'static [usize; 255] {
-        &AMINOACID_INDEX
-    }
-
-    fn get_q(&self) -> &SubstMatrix {
-        &self.q
-    }
-
-    fn get_stationary_distribution(&self) -> &FreqVector {
-        &self.params.pi
-    }
-
-    fn normalise(&mut self) {
-        let factor = -(self.params.get_pi().transpose() * self.q.diagonal())[(0, 0)];
-        self.q /= factor;
-    }
-}
-
-impl EvolutionaryModel for ProteinSubstModel {
-    type ModelType = ProteinModelType;
-    type Params = ProteinSubstParams;
-
-    fn new(model: ProteinModelType, params: &[f64]) -> Result<Self>
-    where
-        Self: Sized,
-    {
-        SubstitutionModel::new(model, params)
-    }
-
-    fn get_p(&self, time: f64) -> SubstMatrix {
-        SubstitutionModel::get_p(self, time)
-    }
-
-    fn get_q(&self) -> &SubstMatrix {
-        SubstitutionModel::get_q(self)
-    }
-
-    fn get_rate(&self, i: u8, j: u8) -> f64 {
-        SubstitutionModel::get_rate(self, i, j)
-    }
-
-    fn get_stationary_distribution(&self) -> &FreqVector {
-        SubstitutionModel::get_stationary_distribution(self)
-    }
-
-    fn get_char_probability(&self, char_encoding: &FreqVector) -> FreqVector {
-        let mut probs = SubstitutionModel::get_stationary_distribution(self)
-            .clone()
-            .component_mul(char_encoding);
-        probs.scale_mut(1.0 / probs.sum());
-        probs
-    }
-
-    fn index(&self) -> &[usize; 255] {
-        SubstitutionModel::index(self)
-    }
-
-    fn get_params(&self) -> &Self::Params {
-        &self.params
-    }
-}
-
-impl ParsimonyModel for ProteinSubstModel {
-    fn generate_scorings(
-        &self,
-        times: &[f64],
-        zero_diag: bool,
-        rounding: &Rounding,
-    ) -> HashMap<OrderedFloat<f64>, (SubstMatrix, f64)> {
-        SubstitutionModel::generate_scorings(self, times, zero_diag, rounding)
-    }
-
-    fn get_scoring_matrix(&self, time: f64, rounding: &Rounding) -> (SubstMatrix, f64) {
-        SubstitutionModel::get_scoring_matrix(self, time, rounding)
-    }
-}
-
-pub type ProteinSubstModelInfo = SubstModelInfo<ProteinSubstModel>;
-pub type ProteinLikelihoodCost<'a> = SubstitutionLikelihoodCost<'a, ProteinSubstModel>;
-
-impl<'a> LikelihoodCostFunction<'a> for ProteinLikelihoodCost<'a> {
-    type Model = ProteinSubstModel;
-    type Info = ProteinSubstModelInfo;
-
-    fn compute_log_likelihood(&self) -> f64 {
-        self.compute_log_likelihood().0
-    }
-
-    fn get_empirical_frequencies(&self) -> FreqVector {
-        let all_counts = self.info.get_counts();
-        let mut total = all_counts.values().sum::<f64>();
-        let index = SubstitutionModel::index(self.model);
-        let mut freqs = frequencies!(&[0.0; Self::Model::N]);
-        for (&char, &count) in all_counts.iter() {
-            freqs += &PROTEIN_SETS[char as usize].scale(count);
-        }
-        for &char in AMINOACIDS {
-            if freqs[index[char as usize]] == 0.0 {
-                freqs[index[char as usize]] += 1.0;
-                total += 1.0;
-            }
-        }
-        freqs.map(|x| x / total)
     }
 }
